Working on more genes and basic AI when I encountered some very grave problems, I hope I can resolve that without having to completely change everything.

+++ src/main/java/seremis/geninfusion/soul/traits/TraitFire.java
@@ -47,7 +47,7 @@ public class TraitFire extends Trait {
                             }
                         }
                     } else {
-                        entity.setPersistentVariable("fire", (int)(8*20));
+                        entity.setPersistentVariable("fire", 8*20);
                     }
                 }
             }
@@ -74,7 +74,7 @@ public class TraitFire extends Trait {
         if(UtilSoulEntity.handleLavaMovement(entity)) {
             if(!isImmuneToFire) {
                 entity.attackEntityFrom(DamageSource.lava, 4.0F);
-                entity.setPersistentVariable("fire", (int)(15*20));
+                entity.setPersistentVariable("fire", 15*20);
             }
             entity.setPersistentVariable("fallDistance", fallDistance * 0.5F);
         }
